implement yaxis stepper and yaxis canopen
